udev: add option to generate old 'buggy' serials
Prior to udev 184, scsi_id would truncate the last character of the model
string when generating the ID_SERIAL value. If a system was installed
prior to that fix being available in udev, there may be configuration
information that refers to the truncated link.

This patch adds a --truncated-serial option and a udev rule will created
the old truncated links.

Signed-off-by: Jeff Mahoney <jeffm@suse.com>