tmpfiles: make create_file() safe
The path leading to the target file is now validated by chase_symlinks() and
the use of CHASE_SAFE flag.

This has several consequences:

 - It will fail on any unsafe transition that might be encountered on the
   target path.

 - Unsafe transitions is not limited to symlink ones: transitions such as
   "directory->directory" or "directory->file" are also checked.

 - Since the validation is done on the dirname part of the path, it's still
   possible to create an object owned by root in a directory owned by a regular
   user.

The file descriptor returned by chase_symlink() is then used by the rest of the
function since it's been converted to operate on fd only.

This can lead to incompatibilities.