journald: rename system_{journal,storage} and runtime_{journal,storage} into persistent_{journal,storage} and volatile_{journal,storage} respectively
The "system" journal is actually the journal containing messages sent by either
system services or by root user and can be either persistent or volatile.

The code was using the term "system_journal" for system journal exclusively in
persistent mode which was a bit misleading because as described previously the
system journal can be either in persistent mode or in volatile mode.

This patch renames the places where "system" where used instead of persistent.

No functional change.