journald: add user journal instance support
Idea was mentioned by Daan De Meyer at
https://github.com/systemd/systemd/pull/22735#issuecomment-1230030823

This allows an unpriviledged user to run his own instance of journald. In this
case user services have stdout/stderr connected to the user instance rather
than the system one, if one is currently running. The log data are saved in
location owned by the user. This should result in a better resource accounting
for a given user and also might decrease the workload of the system instance in
certain cases.

For now all user log sent to the system instance (especially via /dev/log) are
still handled by the system journal and therefore stored in
/var/log/journal. This will be fixed by a later patch though.

Some notes:

- --user can only be used by unprivildged users. IOW root has always its logs
    saved in system.journal.

- user instances' config files are parsed in the following locations (in order
  of precedence):

   - ~/.config/systemd/journald-user.conf
   - /etc/systemd/journald-user.conf

- user instances are optional: if they're disabled, log data of user services
  are handled by the system instance as before.

- a user instance for root is not allowed, root logs are sent to the system
  instance as usual.

- user instance and namespace: log namespaces are not allowed for user services
  hence user instances refuse to start if a namespace name is provided.

- persistent storage of an user instance is always accessible.